refactor: decompose ChatScreen and CatalogScreen monoliths
- chat.py (934→579): extract slash commands into SlashCommandsMixin
  in chat_commands.py
- catalog.py (833→680): extract TableRow and row-builder functions
  into catalog_types.py
