+++ src/lilbee/cli/tui/screens/setup.py
@@ -21,7 +21,7 @@ from lilbee.catalog import (
     make_download_callback,
 )
 from lilbee.cli.tui import messages as msg
-from lilbee.cli.tui.screens.catalog import (
+from lilbee.cli.tui.screens.catalog_types import (
     TableRow,
     catalog_to_row,
     format_size_gb,
